Fixe #43
This fix may be incomplete. When I delete a card, I see the following message in the console:

Uncaught Error in snapshot listener: FirebaseError: [code=permission-denied]: Missing or insufficient permissions.

But the remove feature seems to be working just fine and everything looks fine from the Firebase console. Not sure why we are seeing that error, but for now it seems ok to ignore it.